Refactor: re-export 적용
