+++ docs/REQUIREMNET.md
@@ -64,31 +64,31 @@
 - [x] _당첨 번호와 보너스 번호를 입력 할 수 있다._
   - [x] _입력받은 숫자가 총 6개인가?_
   - [x] _입력값이 모두 1에서 45 사이의 숫자로 이루어져있는가?_
-  - [ ] _입력값 중에 중복된 숫자가 존재하는가?_
+  - [x] _입력값 중에 중복된 숫자가 존재하는가?_
 - [x] _보너스 번호를 입력 할 수 있다._
-  - [ ] _입력값이 1에서 45 사이의 숫자인가?_
-  - [ ] _입력값 중에 중복된 숫자가 존재하는가?_
+  - [x] _입력값이 1에서 45 사이의 숫자인가?_
+  - [x] _입력값 중에 중복된 숫자가 존재하는가?_
 - [x] 구매한 로또의 결과를 알 수 있다.
 - [x] 수익률을 알 수 있다.
 
 ## 프로그램 컨트롤 로직 추가
 
-- [ ] 구매 금액를 입력 할 수 있다.
-  - [ ] 입력값이 구매 단위에 비례하는 숫자인가?
-  - [ ] 구매하는 로또의 갯수가 최소 구매 갯수 이상인가?
-  - [ ] 구매하는 로또의 갯수가 최대 구매 갯수 이하인가?
-- [ ] 로또를 구매 금액에 비례하여 구매할 수 있다.
-- [ ] 로또를 1장 구매 할 수 있다.
-- [ ] 당첨 번호와 보너스 번호를 입력 할 수 있다.
-  - [ ] 입력받은 숫자가 총 6개인가?
-  - [ ] 입력값이 모두 1에서 45 사이의 숫자로 이루어져있는가?
-  - [ ] 입력값 중에 중복된 숫자가 존재하는가?
-- [ ] 보너스 번호를 입력 할 수 있다.
-  - [ ] 입력값이 1에서 45 사이의 숫자인가?
-  - [ ] 입력값 중에 중복된 숫자가 존재하는가?
-- [ ] 구매한 로또의 결과를 알 수 있다.
-- [ ] 수익률을 알 수 있다.
-- [ ] _프로그램을 종료한다._
+- [x] 구매 금액를 입력 할 수 있다.
+  - [x] 입력값이 구매 단위에 비례하는 숫자인가?
+  - [x] 구매하는 로또의 갯수가 최소 구매 갯수 이상인가?
+  - [x] 구매하는 로또의 갯수가 최대 구매 갯수 이하인가?
+- [x] 로또를 구매 금액에 비례하여 구매할 수 있다.
+- [x] 로또를 1장 구매 할 수 있다.
+- [x] 당첨 번호와 보너스 번호를 입력 할 수 있다.
+  - [x] 입력받은 숫자가 총 6개인가?
+  - [x] 입력값이 모두 1에서 45 사이의 숫자로 이루어져있는가?
+  - [x] 입력값 중에 중복된 숫자가 존재하는가?
+- [x] 보너스 번호를 입력 할 수 있다.
+  - [x] 입력값이 1에서 45 사이의 숫자인가?
+  - [x] 입력값 중에 중복된 숫자가 존재하는가?
+- [x] 구매한 로또의 결과를 알 수 있다.
+- [x] 수익률을 알 수 있다.
+- [x] _프로그램을 종료한다._
 
 # 어플리케이션 확장 예상해보기
