+++ src/js/domain/LottoMachine.js
@@ -1,14 +1,14 @@
-import { LOTTO_NUMBER_QUANTITY, LOTTO_PRICE } from '../constants/lotto-config.js';
+import { LOTTO_NUMBER_QUANTITY, LOTTO_NUMBER_RANGE, LOTTO_PRICE } from '../constants/lotto-config.js';
 import getRandomNumber from '../utils/getRandomNumber.js';
-import checkValidPurchase from '../validator/purchase.js';
+import { checkValidPurchase } from '../validator/index.js';
 import Lotto from './Lotto.js';
 
 class LottoMachine {
   #numberQuantity = LOTTO_NUMBER_QUANTITY;
 
-  #minNumber = 1;
+  #minNumber = LOTTO_NUMBER_RANGE.MIN;
 
-  #maxNumber = 45;
+  #maxNumber = LOTTO_NUMBER_RANGE.MAX;
 
   #pricePerSheet = LOTTO_PRICE;
