+++ src/js/validator/index.js
@@ -1,0 +1,3 @@
+export { default as checkValidBonus } from './bonus.js';
+export { default as checkValidPurchase } from './purchase.js';
+export { default as checkValidWinningNumbers } from './winningNumbers.js';
